feat: refactoring create_configmap_or_secret to kubernetes_manager module
