+++ app/v1/modules/kubernetes_manager/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 
+from .script_create_configmap_or_secret import ScriptConfigMapSecretCreator
 from .script_create_environment_values_file import ScriptEnvironValuesFileCreator
 from .script_create_vars_file import ScriptVarsFileCreator
 from .script_create_workload_template import ScriptWorkloadTemplateCreator
@@ -8,6 +9,7 @@ from .script_create_workload import ScriptWorkloadCreator
 
 
 __all__ = [
+    "ScriptConfigMapSecretCreator",
     "ScriptEnvironValuesFileCreator",
     "ScriptVarsFileCreator",
     "ScriptWorkloadTemplateCreator",
